feature: Serialize number, bool, string types
