+++ Cargo.toml
@@ -21,3 +21,7 @@ unstable_parse = []
 [dependencies]
 serde = { version = "1.0.*", optional = true }
 ordered-float = "3.4.0"
+
+[dev-dependencies]
+serde = "1.0.152"
+serde_derive = "1.0.152"
